requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.19.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.1
idna==3.6
motor==3.3.2
multidict==6.0.4
pymongo==4.6.1
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
urllib3==2.1.0
yarl==1.9.4
